+++ src/index.js
@@ -1,7 +1,10 @@
 /*eslint-disable*/
 import _ from 'lodash';
 import './style.css';
-import render from './modules/display.js';
+// import render from './modules/display.js';
+import { viewShows } from './modules/display.js';
+viewShows();
 
-window.addEventListener('DOMContentLoaded', render);
+
+//  window.addEventListener('DOMContentLoaded', render);
